+++ game_patch/rf/geometry.h
@@ -375,5 +375,23 @@ namespace rf
     };
     static_assert(sizeof(GProceduralTexture) == 0x38);
 
+    struct GPortalObject
+    {
+        unsigned int id;
+        Vector3 pos;
+        float radius;
+        bool has_alpha;
+        bool did_draw;
+        bool is_behind_brush;
+        bool lights_enabled;
+        bool use_static_lights;
+        Plane *object_plane;
+        Vector3 *bbox_min;
+        Vector3 *bbox_max;
+        float z_value;
+        void (*render_function)(int, GSolid *);
+    };
+    static_assert(sizeof(GPortalObject) == 0x30);
+
     static auto& g_cache_clear = addr_as_ref<void()>(0x004F0B90);
 }